requirements.txt
# GEPA and DSPy dependencies for payment optimization evolution
dspy-ai>=2.4.0
gepa>=0.1.0
pandas>=1.5.0
numpy>=1.24.0
openai>=1.0.0
perplexity-ai>=0.1.0

# Optional: For advanced evolution features
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# For API integration
requests>=2.31.0
python-dotenv>=1.0.0
